backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pymongo[srv]==4.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==1.10.13
python-multipart==0.0.9
